Add changes after ACE v3.16.0 release
Includes a large amount of changes, especially to repair. The repair
settings will probably still require some tweaking, since ACE switched
from a constant Full Repair time to a dynamic one, which seems to
fluctuate *a lot* based on the number of repairable parts of a vehicle.
Advanced Vehicle Damage was disabled because upstream changed the
default value to disabled, plus I'm suspecting that it might contribute
to the long repair times.

Additionally, added some TODO entries for checking some of the settings
in the future, as well as some clarifying comments.

The most notable Medical addition is grave digging, which has been
disabled for now, at least.